// src/api/generated/model/getProductBySlug200VariationTypes.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Nero API
 * A Nero API é uma interface RESTful robusta projetada especificamente para o varejo de moda online. Ela oferece controle total sobre o ciclo de vida do produto, desde a entrada de estoque em múltiplas variantes (SKUs) até o processamento final do checkout. Ideal para marketplaces, lojas boutique ou aplicativos móveis de moda.
 * OpenAPI spec version: 1.0.0
 */

export type GetProductBySlug200VariationTypes = ({
  id: string;
  name: string;
  position: number;
  hasImage: boolean;
  options: ({
  id: string;
  value: string;
  imageUrl: string | null;
  position: number;
})[];
})[] | null;